Added basic webpack build process
