+++ .gitignore
@@ -1,2 +1,3 @@
 .sass-cache/
 *.css.map
+node_modules/
