+++ bands.html
@@ -4,8 +4,8 @@
     <link rel="shortcut icon" type="image/png" href="images/favico.png"/>
     <link rel="stylesheet" type="text/css" href="stylesheets/style.css">
     <link href='https://fonts.googleapis.com/css?family=Amatic+SC&subset=latin,latin-ext' rel='stylesheet' type='text/css'>
-    <script type="text/javascript" src="javascripts/jquery-2.0.3.min.js"></script>
-    <script type="text/javascript" src="javascripts/index.js"></script>
+    <script type="text/javascript" src="javascript/jquery-2.0.3.min.js"></script>
+    <script type="text/javascript" src="javascript/index.js"></script>
 </head>
 <body>
     <div class="stars"></div>
