+++ game.html
@@ -3,10 +3,10 @@
 <head>
 	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 	<title>co-letter-all [BETA]</title>
-	<script src="javascripts/jquery-2.0.3.min.js"></script>
-	<script src="javascripts/game-script.js"></script>
+	<script src="javascript/jquery-2.0.3.min.js"></script>
+	<script src="javascript/game-script.js"></script>
 	<!-- script file with word lists and alphabet array -->
-	<script src="javascripts/words.js"></script>
+	<script src="javascript/words.js"></script>
 	<link rel="stylesheet" type="text/css" href="stylesheets/stil.css"/>
 	<link href='http://fonts.googleapis.com/css?family=Press+Start+2P&subset=latin,latin-ext' rel='stylesheet' type='text/css'>
 	<link href='https://fonts.googleapis.com/css?family=Press+Start+2P&subset=latin,latin-ext' rel='stylesheet' type='text/css'>
