+++ index.html
@@ -11,8 +11,8 @@
     <link rel="stylesheet" type="text/css" href="stylesheets/style.css">
     <link href='https://fonts.googleapis.com/css?family=Amatic+SC&subset=latin,latin-ext' rel='stylesheet' type='text/css'>
     <script src="https://kit.fontawesome.com/4973abd21e.js" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="javascripts/particles.min.js"></script>
-    <script type="text/javascript" src="javascripts/index.min.js" defer></script>
+    <script type="text/javascript" src="javascript/particles.min.js"></script>
+    <script type="text/javascript" src="javascript/index.min.js" defer></script>
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-S9273E2R33"></script>
     <script>
