+++ spinster.html
@@ -10,8 +10,8 @@
         <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp" crossorigin="anonymous">
         <link href='https://fonts.googleapis.com/css?family=Roboto+Condensed' rel='stylesheet' type='text/css'>
         <link href='https://fonts.googleapis.com/css?family=Slabo+27px&subset=latin,latin-ext' rel='stylesheet' type='text/css'>
-        <script src="javascripts/jquery-2.0.3.min.js"></script>
-        <script src="javascripts/spinster.js"></script>
+        <script src="javascript/jquery-2.0.3.min.js"></script>
+        <script src="javascript/spinster.js"></script>
     </head>
     <body>
         <div class="mainContainer">
